MAINT: Variadic parameters in template template.
The purpose of this code is to be parametric over the choice of either
`std::shared_ptr` or `std::unique_ptr`. The previous code assumed that
both shared_ptr and unique_ptr were single-argument templates, but
unique_ptr has a second template parameter, so clang complains because
it doesn't think unique_ptr is a valid substitution for the template
template param.