BUG: fix compilation on GCC 7 and improve error compiler errors
Something about the template variadic value parameter auto type inferencer in
GCC 7 appears to be broken which causes the constructor for (at least) row view
to be broken. The issue appears to be as follows:

template<auto... values>
struct s;

using T = s<reinterpret_cast<int*>(nullptr),
            reinterpret_cast<double*>(nullptr),
            reinterpret_cast<char*>(nullptr)>;

In this example, it seems that all the values get inferred to be of type int*
because that is the type of the first argument. It then reinterpret casts the
pointers around silently losing the type information. It seems that the fact
that the pointers are initialized to nullptr is required to reproduce this
issue.

This change uses a new template variable named column_singleton to serve as the
object whose address is being taken to produce a pointer value to pass to table,
table_view, row, etc. Using a template variable instead of a static member of
the column class solves the storage declaration problem so there are no linker
errors even when compiling with optlevel 0.