ENH: add any_ref, any_cref and ndarray_views of these types
The any_ref type is a reference-like object whose type isn't known
statically. The any_ref object holds onto a value which represents the type at
runtime, so it can still be safely cast back to a static value.

The any_cref type is a const reference-like object whose type isn't known
statically. Unlike any_ref, any_cref cannot mutate the contents of the
referent, and may only be statically cast back to a const object.

The ndarray_view specializations for these types only hold one copy of the
runtime type information on the array itself. When indexing into the array or
iterating the array, an any_ref or any_cref will be created on the fly to
represent the untyped memory at the given index. The entire array can be cast
back to a statically typed ndarray_view.