BUG: define to_object(datetime64) before we create the vtable
Moves the definition of `py::dispatch::to_object<datetime<unit>>` before
`dtype_to_vtable` and `vtable_to_dtype` because these function instantiate
vtables which conditionally implement the `to_object` slot based on the result
of `py::has_to_object`. If we create the vtable before `to_object(datetime64)`
is defined, the check is False so we generate a method that throws an exception.