BUG: normalize exception before reading the value
Some exceptions are not actually allocated until required to make it faster to
throw and catch exceptions inside the interpreter. There are a few cases I know
of where this happens:

- GeneratorExit being raised in a generator as the result of gen.close()
- KeyboardInterrupt being raised as the result of C-c

I caught this by having a C++ extension that called back into Python code, which
raised a KeyboardInterrupt, which, when we went to throw it back into Python,
wasn't initialized. The `type` was `PyExc_KeyboardInterrupt`, but the `value`
field was just `nullptr`.