ENH: add no destruct wrapper for static caches
Sometimes the interpreter clears out objects in caches before the static
variable destructors get invoked. This can lead to a double-free when a
`scoped_ref` is deleted for the second time.

This change adds a `py::detail::no_destruct_wrapper` which allocates space using
a `std::aligned_storage_t` and placement new's the object into this buffer. This
allow us to skip destruction of the variable. The object is accessed through
`get()` member variables which return `T&` or `const T&`.

This is currently in `py::detail`, but may be promoted to `py::util` if we find
this is useful in applications.