BUG: Invoke import_array as a macro rather than a function.
When upgrading libpy in qexec, we say failures where the
ensure_import_array_module_scope constructor was invoked, but Py_Array_API was
NULL, which suggests that the function-based version of our import_array
automation code wasn't working properly. Moving it to a macro, which causes
import_array to be invoked "directly" inline in the scope of the macro, appears
to fix the issue, though I still don't totally understand why.

This also adds a missing invocation of import_array() in csv_writer.cc.