DEV: New build system and factor out csv support.
Adds pip packaging for libpy to make it easier to build libraries that use
libpy. The pip package includes Python support code for loading libpy into a
program and for building extension modules that use libpy.

The CSV parser and writer are pulled into a standalone library, currently called
`qcsv` (but subject to change). The CSV parser is not particularly general
purpose, and require some Python support code, which makes it a bad candidate
for keeping in libpy itself. The other reason to pull the CSV support out of
libpy is to create an example library that consumes libpy.

To use the Python support library to build an extension with libpy, you need to
install the `libpy` package before running your `setup.py`. This is sort of
unfortunate, but attempts to automatically do this for numpy in other projects
have been far more trouble than they are worth. Below is a sample of how you
would use libpy in a setup.py:

```python
from libpy.build import LibpyExtension
from setuptools import setup

setup(
    # ...
    name='my_module',
    ext_modules=[
        LibpyExtension(
            'my_module._extension',
            ['my_module/_extension.cc'],
        ),
    ],
    install_requires=['libpy'],
)
```

This will set up the libpy include paths and everything for your extension.
The only other thing you need to do is put the following in your
`PyMODINIT_FUNC`:

```c++
if (py::abi::ensure_compatible_libpy_abi()) {
    return nullptr;  // or just return in Python 2
}
```

This check will ensure the currently installed of libpy is still compatible with
your extension module. This check is needed because installing a new library may
upgrade libpy without you realizing it. This check will protect your module from
any weird ABI incompatibilities at runtime.

The other thing that you need to do is to ensure that `libpy` has been imported
in Python before attempting to import any of the extension modules that use
libpy. This should be done by importing `libpy` in the top level `__init__.py`
of package that has any libpy extensions.