ENH: Add shorter names for py::arg::* types.
`py::arg::*` types are useful for can make it hard to read a type signature.
Given the the namespace `py::arg`, shorter names are not harder to read and
won't likely clash with other unrelated features.

This change also accumulates the num optional arguments in a `constexpr size_t`
instead of maintaining a tuple of optional args and then only using the
`std::tuple_size_v` of the tuple. This should reduce the number of types created
during compilation.