BLD: Proper DSO visibility rules.
Shared objects on GNU+Linux and macOS export all of their symbols by default.
This can cause ODR violations when multiple versions of libpy are used and
inline functions are put into more than one extension module at once. Each
extension module will export their own version of the inline function which can
cause problems if the definitions differ. To make the ABI versioning of libpy
easier, we now compile with `-fvisibility=hidden` to make symbols hidden by
default. Anything that wants to be exported should be marked with `LIBPY_EXPORTED`
to indicate that this is part of the public libpy API.

As a consumer of libpy you can build extensions that libpy in two ways:

1. Using the `libpy.build.LibpyExtension` type and setuptools
1. Some custom build system.

If you are using the setuptools build, you won't need to change anything about
your code.

If you are building with a custom build system, you will need to pass
`-fvisibility-inlines-hidden` to not re-export inlined functions defined in
libpy. If you are already passing `-fvisibility=hidden` then you don't need to
add this flag, as `-fvisibility-inlines-hidden` is a weaker flag than
`-fvisibility=hidden`.

If you are modifying libpy, you may need to do some extra work. To decide if
you need to add one of these markers, you may follow this guide:

- are you adding a new template type or template function?
  - no extra code needed
- are you adding a new inline function?
  - consider putting the implementation in a `.cc` file and check this
    guide again
  - otherwise, no extra code needed
- are you adding a new template specialization for an existing type?
  - no extra code needed
- are you adding a new function that is only used within a single `.cc` file as
  an implementation detail?
  - no extra code needed
- are you adding a new function which is a public API function whose
  implementation is in a `.cc` file?
  - prefix the function declaration (in the `.h` file) with `LIBPY_PUBLIC`
- are you adding a new type which is a public API type whose member functions
  are implemented in a `.cc` file?
  - put `LIBPY_BEGIN_EXPORT` before the beginning of the type declaration
    (in the `.h` file)
  - put `LIBPY_END_EXPORT` after the end of the type declaration (in the `.h`
    file)