BUG: fix early reference invalidation in any_vector::push_back
While `vector.push_back` invalidates references,
`vector.push_back(vector.front())` should still work. For that to happen, we
must copy from the new reference before we invalidate the object. This changes
`any_vector::grow` to take the value to insert such that we can safely copy the
value before destroying the old one.
The reference must also be copied before the objects have been _moved_ from the
old buffer into the new one, otherwise we will copy the moved-from state of the
input argument which is not the desired behavior.

Finally, this change fixes a bug which treated
`any_vector::push_back(py::any_ref&&)` like `any_vector::push_back(T&&)` where `T`
is the static type of the vector's members. This means that we would _move_ from
the reference, even though it was not expiring. This created a problem for code
like:

```c++
void f(py::any_vector& v) {
    v.push_back(v.front());
}
```

because `v.front()` returned a new `py::any_ref` object which was considered an
expiring reference in this context. The same bug was not present given code like:

```c++
void f(py::any_vector& v) {
    auto front_ref = v.front();
    v.push_back(front_ref);
}
```

because not `front_ref` is passed as a `const py::any_ref&` instead of a
`py::any_ref&&`. This all applies for `py::any_cref` as well, though it was
harder to form this construction.