ENH: optimized access to contiguous ndarray_views
When iterating an `ndarray_view` which is actually contiguous, there is
performance penalty over iterating the underlying pointer. This is because the
compiler cannot optimize the loop as well when each step is adding a
non-constant value. This will also defeat any vectorization efforts from the
compiler.

To help get this performance back, 3 methods were added to `ndarray_view`:

- `is_c_contig()`
- `is_f_contig()`
- `is_contig()`

These methods check the contiguity of the array to allow a programmer to use a
more efficient function when the contiguity is known.

There is also a new free function `for_each_unordered` which takes an
`ndarra_view` and a unary function. The function is applied to each element in
the array in an unspecified order. The order of application is left unspecified
to allow selecting an optimal traversal. This function is also convenient for
touching all of the elements of multidimensional array, which don't currently
support iteration.