BUG: add missing type check to from_object<T&>(py::borrowed_ref)`
`from_object<T&>(py::borrowed_ref)` is designed to pull out a reference to a
type T which has been adapted to Python with `py::autoclass`. This acts as a
type-safe interface over `unbox`. Currently, the code just checks that `T` has
been registered with autoclass; however, ignores the type check that ensures
that the passed object is an instance of the given type. This adds the type
check for `PyObject_IsInstance` against the runtime value and the registered
Python type for `T`.