BUG: fix `py::from_object<T&>` with non-default autoclass base
When an autoclass uses a non-default base, which is usually used to create
subclasses of other Python types, like `PyExceptionBaseObject`, we need to case
the `PyObject*` passed to `py::from_object` to the correct static type. The
correct type should be: `py::autoclass<T, base>::object`, but we are currently
just casting to `py::autoclass<T>::object`, which is an unrelated type. This
means we are getting the wrong address for the value being wrapped in Python
object.

To fix this, the autoclass type cache now stores an
`unbox_fn` (`void* (*)(py::borrowed_ref<>)`) member which is a function that can
properly unbox a `PyObject`. This function is initialized from the original
`py::autoclass` struct specialization that knows about both `T` and `base`.
`py::from_object` now uses this function from the cache entry that it looked up
to do the `PyObject_IsInstance` check.

`unbox_fn` returns a `void*` because it must have a uniform return type across
all of the entries in the autoclass cache. The contract is that the function
gets the proper pointer to the contained object and returns it as a void
pointer. Consumers can then cast the pointer back to a `T*` for the known `T`.